requirements.txt
click>=8.0
ring-attention-pytorch
pybind11>=2.11
ninja
pytest